Added a wrapper function for sign flipping.
