+++ osl/sign_flipping/__init__.py
@@ -2,4 +2,4 @@
 
 """
 
-from .sign_flipping import *  # noqa: F401, F403
+from .wrappers import *  # noqa: F401, F403
